+++ js/valid_check.js
@@ -1,0 +1,5 @@
+import './io_script.js';
+import {checkCommentMax} from './utils.js';
+
+checkCommentMax();
+
